@@ -4,22 +4,21 @@
   "private": true,
   "homepage": "https://jacobrec.github.io/biketrip-viewer/",
   "dependencies": {
-    "@emotion/react": "^11.4.1",
-    "@emotion/styled": "^11.3.0",
-    "@material-ui/icons": "^4.11.2",
-    "@mui/icons-material": "^5.0.3",
-    "@mui/material": "^5.0.3",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
-    "@types/mapbox-gl": "^2.4.2",
-    "d3": "^7.1.1",
-    "mapbox-gl": "^2.5.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-scripts": "4.0.3",
-    "typescript": "^4.4.3",
-    "web-vitals": "^1.1.2",
+    "@emotion/react": "^11.11.3",
+    "@emotion/styled": "^11.11.0",
+    "@mui/icons-material": "^5.15.10",
+    "@mui/material": "^5.15.10",
+    "@testing-library/jest-dom": "^6.4.2",
+    "@testing-library/react": "^14.2.1",
+    "@testing-library/user-event": "^14.5.2",
+    "@types/mapbox-gl": "^2.7.21",
+    "d3": "^7.8.5",
+    "mapbox-gl": "^3.1.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-scripts": "^5.0.1",
+    "typescript": "^4.9.5",
+    "web-vitals": "^3.5.2",
     "worker-loader": "^3.0.8"
   },
   "scripts": {
@@ -53,8 +52,8 @@
     ]
   },
   "devDependencies": {
-    "@types/d3": "^7.0.0",
-    "@types/react": "^17.0.27",
-    "gh-pages": "^3.2.3"
+    "@types/d3": "^7.4.3",
+    "@types/react": "^18.2.57",
+    "gh-pages": "^6.1.1"
   }
 }
