@@ -9,7 +9,8 @@ import * as d3 from "d3";
 // eslint-disable-next-line import/no-webpack-loader-syntax
 (mapboxgl as any).workerClass = require('worker-loader!mapbox-gl/dist/mapbox-gl-csp-worker').default;
 
-mapboxgl.accessToken = 'pk.eyJ1Ijoic3VqYW5jaGFrcmFib3J0eSIsImEiOiJja2Q5MzBuc3owenplMnBzY2I0eDYwdDhvIn0.imItePLDlYNF2BGVde_mkw';
+// This token is locked to the URL: https://jacobrec.github.io/biketrip-viewer/
+mapboxgl.accessToken = 'pk.eyJ1IjoicmVja2hhcmQiLCJhIjoiY2xzdXR6eXc0MXl2MTJrbnc2djgzbHI5bSJ9.4boUfw73i1AnAoRTAdY9uQ';
 
 type DataInfo = {
   locations: string[],
